+++ packages/admin/src/themeAugmentation.ts
@@ -7,7 +7,6 @@ import {
     CometAdminInputBaseClassKeys,
     FieldContainerThemeProps,
 } from "./form";
-import { CometAdminFixedLeftRightLayoutKeys } from "./layout";
 import {
     CometAdminBreadcrumbsClassKeys,
     CometAdminBreadcrumbsThemeProps,
@@ -32,7 +31,6 @@ declare module "@material-ui/core/styles/overrides" {
         CometAdminErrorBoundary: CometAdminErrorBoundaryClassKeys;
         CometAdminFormPaper: CometAdminFormPaperKeys;
         CometAdminFormSection: CometAdminFormSectionKeys;
-        CometAdminFixedLeftRightLayout: CometAdminFixedLeftRightLayoutKeys;
         CometAdminBreadcrumbs: CometAdminBreadcrumbsClassKeys;
     }
 }